+++ kernel/src/Memory/kmalloc.hpp
@@ -6,7 +6,7 @@
 void kmalloc_init();
 
 void* kmalloc(size_t size);
-void* krealloc(void* addr, size_t size);
+void* kcalloc(size_t size);
 void kfree(void* ptr);
 
 #endif /* _KERNEL_MALLOC_HPP */